src/main/webapp/app/entities/estrangeiro/delete/estrangeiro-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { EstrangeiroService } from '../service/estrangeiro.service';

import { EstrangeiroDeleteDialogComponent } from './estrangeiro-delete-dialog.component';

describe('Estrangeiro Management Delete Component', () => {
  let comp: EstrangeiroDeleteDialogComponent;
  let fixture: ComponentFixture<EstrangeiroDeleteDialogComponent>;
  let service: EstrangeiroService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [EstrangeiroDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(EstrangeiroDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(EstrangeiroDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(EstrangeiroService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
